+++ src/lib/components/consignments/__stories__/index.tsx
@@ -1,13 +1,13 @@
 import "./consignments.story"
-import "./typography.story"
+import "./style.story"
 
 import { storiesOf } from "@storybook/react-native"
 import * as React from "react"
 import { View } from "react-native"
 
-import * as bottomAlignedButton from "./consignments-bottom-aligned.story"
-import * as search from "./consignments-search.story"
-import * as todo from "./consignments-todo.story"
+import * as bottomAlignedButton from "./bottom_aligned.story"
+import * as search from "./search.story"
+import * as todo from "./todo.story"
 
 // Just a quick interface so you know the API
 export interface AutoStory {
@@ -15,11 +15,10 @@ export interface AutoStory {
   allStates: any[]
   component: any
 }
-
 // Converts a set of state and a component type into a set of stories
 const autoStories: AutoStory[] = [search, bottomAlignedButton, todo]
 
-const Wrapper = props =>
+export const Wrapper = props =>
   <View style={{ flex: 1, backgroundColor: "black", padding: 20, marginTop: 60 }}>
     {props.children}
   </View>